+++ src/main/java/id/xtramile/indonesia/IndonesiaDataLoader.java
@@ -7,10 +7,10 @@ import java.util.Map;
 
 public interface IndonesiaDataLoader {
 
-    Map<Integer, Province> loadProvinces() throws DataLoadException;
-    Map<Integer, City> loadCities() throws DataLoadException;
-    Map<Integer, District> loadDistricts() throws DataLoadException;
-    Map<Integer, Village> loadVillages() throws DataLoadException;
+    Map<Long, Province> loadProvinces() throws DataLoadException;
+    Map<Long, City> loadCities() throws DataLoadException;
+    Map<Long, District> loadDistricts() throws DataLoadException;
+    Map<Long, Village> loadVillages() throws DataLoadException;
 
     IndonesiaData loadAllData() throws DataLoadException;
 }
