+++ src/main/java/id/xtramile/indonesia/model/IndonesiaData.java
@@ -3,31 +3,31 @@ package id.xtramile.indonesia.model;
 import java.util.Map;
 
 public class IndonesiaData {
-    private final Map<Integer, Province> provinces;
-    private final Map<Integer, City> cities;
-    private final Map<Integer, District> districts;
-    private final Map<Integer, Village> villages;
+    private final Map<Long, Province> provinces;
+    private final Map<Long, City> cities;
+    private final Map<Long, District> districts;
+    private final Map<Long, Village> villages;
 
-    public IndonesiaData(Map<Integer, Province> provinces, Map<Integer, City> cities, Map<Integer, District> districts, Map<Integer, Village> villages) {
+    public IndonesiaData(Map<Long, Province> provinces, Map<Long, City> cities, Map<Long, District> districts, Map<Long, Village> villages) {
         this.provinces = provinces;
         this.cities = cities;
         this.districts = districts;
         this.villages = villages;
     }
 
-    public Map<Integer, Province> getProvinces() {
+    public Map<Long, Province> getProvinces() {
         return provinces;
     }
 
-    public Map<Integer, City> getCities() {
+    public Map<Long, City> getCities() {
         return cities;
     }
 
-    public Map<Integer, District> getDistricts() {
+    public Map<Long, District> getDistricts() {
         return districts;
     }
 
-    public Map<Integer, Village> getVillages() {
+    public Map<Long, Village> getVillages() {
         return villages;
     }
 }
